// src/app/models/student.model.ts
export class Students{

    _id: string='';
    username: string='';
    email: string='';
    password: string='';
    admissionno: string='';
    roles: any;
    fullname:  String='';
    mobile: String='';
    address: String='';
    standard: String='';
    section: String='';
    profile: String='';
    photourl: String='';
    signature: String='';
    subject: String='';
    mysections: String='';
    code: String='';
 
}